added normalization for df
